requirements.txt
plaid-python
pandas
pandas-gbq
google-cloud-bigquery
pyarrow
python-dotenv
pytest